requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.22
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
pytz==2023.3.post1
requests==2.27.1
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.17
